Убирает код из main в thumbnails, добавляет код для закрытия попапов через Esc
