+++ js/util.js
@@ -16,8 +16,10 @@ const getUrl = (folder, format) => function (title) {
 };
 
 const isEscapeKey = (evt) => evt.code === 'Escape';
+
 const isNotFormInput = () => document.activeElement.tagName !== 'INPUT' && document.activeElement.tagName !== 'TEXTAREA';
 
+const hasAllowedTagName = (element, tagNamesArray) => tagNamesArray.includes(element.tagName);
 
 export {
   getRandomNumber,
@@ -26,5 +28,6 @@ export {
   getId,
   getUrl,
   isEscapeKey,
-  isNotFormInput
+  isNotFormInput,
+  hasAllowedTagName
 };
